+++ server/src/main/java/com/onetool/server/api/member/controller/MemberController.java
@@ -62,36 +62,6 @@ public class MemberController {
         return ApiResponse.of(SuccessCode.CREATED, response);
     }
 
-    @PostMapping("/email")
-    public ApiResponse<?> findEmail(@RequestBody MemberFindEmailRequest request) {
-        String email = memberService.findEmail(request);
-        return ApiResponse.onSuccess(email);
-    }
-
-    @PostMapping("/emails/verification-requests")
-    public ApiResponse<?> sendMessage(@RequestParam("email") @Valid String email) {
-        memberService.sendCodeToEmail(email);
-        return ApiResponse.onSuccess("이메일이 발송되었습니다.");
-    }
-
-    @GetMapping("/emails/verifications")
-    public ApiResponse<?> verificationEmail(@RequestParam("email") @Valid String email,
-                                            @RequestParam("code") String authCode) {
-        boolean response = memberService.verifiedCode(email, authCode);
-        return (response) ? ApiResponse.onSuccess("이메일이 인증되었습니다.")
-                : ApiResponse.onFailure("404", "코드가 일치하지 않습니다.", null);
-    }
-
-    @PostMapping("/password")
-    public ApiResponse<?> findPwdCheck(@RequestBody MemberFindPwdRequest request) {
-        boolean successFlag = memberService.findLostPwd(request);
-        if (successFlag) {
-            return ApiResponse.onSuccess("이메일을 발송했습니다.");
-        } else {
-            return ApiResponse.onFailure("403", "이메일 발송 과정에서 오류가 발생했습니다.", null);
-        }
-    }
-
     @PatchMapping
     public ApiResponse<?> updateMember(
             @Valid @RequestBody MemberUpdateRequest request,
@@ -139,19 +109,6 @@ public class MemberController {
         } catch (RuntimeException e) {
             return ApiResponse.onFailure("404", "회원을 찾을 수 없습니다.", null);
         }
-    }
-
-    // TODO : uri 수정 필요
-    @GetMapping("/myPage/myQna")
-    public ApiResponse<List<QnaBoardResponse.QnaBoardBriefResponse>> getMyQna(@AuthenticationPrincipal PrincipalDetails principalDetails) {
-        return ApiResponse.onSuccess(memberService.findQnaWrittenById(principalDetails.getContext()));
-    }
-
-    @GetMapping("/myPurchase")
-    public ApiResponse<List<BlueprintDownloadResponse>> getMyPurchases(@AuthenticationPrincipal PrincipalDetails principalDetails) {
-        Long userId = principalDetails.getContext().getId();
-        List<BlueprintDownloadResponse> blueprints = memberService.getPurchasedBlueprints(userId);
-        return ApiResponse.onSuccess(blueprints);
     }
 
     private ResponseCookie createRefreshTokenCookie(String refreshToken) {
